+++ submission/src/epilot_submission/sdkconfiguration.py
@@ -26,9 +26,9 @@ class SDKConfiguration:
     server_idx: Optional[int] = 0
     language: str = "python"
     openapi_doc_version: str = "1.0.0"
-    sdk_version: str = "1.12.2"
-    gen_version: str = "2.484.0"
-    user_agent: str = "speakeasy-sdk/python 1.12.2 2.484.0 1.0.0 epilot-submission"
+    sdk_version: str = "1.12.3"
+    gen_version: str = "2.486.1"
+    user_agent: str = "speakeasy-sdk/python 1.12.3 2.486.1 1.0.0 epilot-submission"
     retry_config: OptionalNullable[RetryConfig] = Field(default_factory=lambda: UNSET)
     timeout_ms: Optional[int] = None